Add Go theme system with hybrid rendering support
Implemented a hybrid Go/TypeScript theme system that allows themes to work
seamlessly across both languages. Go handles metadata and HTML rendering,
while TypeScript handles async asset loading and image hydration.

New Go Files:
- protos/weewar/v1/themes.proto: Proto definitions for shared types
  * ThemeInfo, ThemeManifest, UnitMapping, TerrainMapping
  * Type-safe across Go and TypeScript

- web/assets/embed.go: Embeds mapping.json files at compile time
  * Uses embed.FS for zero runtime I/O
  * Loads fantasy and modern theme manifests

- web/assets/themes/themes.go: Core interfaces
  * Theme interface: Lightweight metadata operations
  * ThemeAssets interface: Heavy I/O operations (for future Phase 2)

- web/assets/themes/base.go: BaseTheme implementation
  * Loads from proto ThemeManifest
  * Shared by SVG-based themes

- web/assets/themes/default.go: DefaultTheme
  * PNG-based original v1 assets
  * No manifest needed, hardcoded paths

- web/assets/themes/fantasy.go: FantasyTheme
  * SVG-based medieval theme
  * Loads from embedded mapping.json

- web/assets/themes/modern.go: ModernTheme
  * SVG-based military theme
  * Loads from embedded mapping.json

- web/assets/themes/registry.go: Theme factory
  * CreateTheme() for runtime theme switching

New TypeScript Files:
- web/src/ThemeUtils.ts: Static hydration utility
  * hydrateThemeImages() finds .theme-unit-image and .theme-tile-image
  * Reads data attributes and calls theme.setUnitImage/setTileImage
  * Async with Promise.all for all images
  * Reusable across all panels

Modified Files:
- web/src/UnitStatsPanel.ts: Added image hydration
  * Imported ThemeUtils
  * Added async hydrateThemeImages() method

- web/templates/templates.go: Template management
  * Handles .templar.html file loading and rendering

Key Design:
- Theme vs ThemeAssets interface split
- Embedded assets via embed.FS
- Shared mapping.json as single source of truth
- Image hydration pattern with data attributes
- No breaking changes to existing TypeScript code

+++ web/assets/themes/base.go
@@ -17,59 +17,59 @@ func NewBaseTheme(manifest *v1.ThemeManifest) *BaseTheme {
 	}
 }
 
-func (b *BaseTheme) GetUnitName(unitId int) string {
-	if mapping, ok := b.manifest.Units[int32(unitId)]; ok {
+func (b *BaseTheme) GetUnitName(unitId int32) string {
+	if mapping, ok := b.manifest.Units[unitId]; ok {
 		return mapping.Name
 	}
 	return ""
 }
 
-func (b *BaseTheme) GetTerrainName(terrainId int) string {
-	if mapping, ok := b.manifest.Terrains[int32(terrainId)]; ok {
+func (b *BaseTheme) GetTerrainName(terrainId int32) string {
+	if mapping, ok := b.manifest.Terrains[terrainId]; ok {
 		return mapping.Name
 	}
 	return ""
 }
 
-func (b *BaseTheme) GetUnitDescription(unitId int) string {
-	if mapping, ok := b.manifest.Units[int32(unitId)]; ok {
+func (b *BaseTheme) GetUnitDescription(unitId int32) string {
+	if mapping, ok := b.manifest.Units[unitId]; ok {
 		return mapping.Description
 	}
 	return ""
 }
 
-func (b *BaseTheme) GetTerrainDescription(terrainId int) string {
-	if mapping, ok := b.manifest.Terrains[int32(terrainId)]; ok {
+func (b *BaseTheme) GetTerrainDescription(terrainId int32) string {
+	if mapping, ok := b.manifest.Terrains[terrainId]; ok {
 		return mapping.Description
 	}
 	return ""
 }
 
-func (b *BaseTheme) GetUnitPath(unitId int) string {
-	if mapping, ok := b.manifest.Units[int32(unitId)]; ok {
+func (b *BaseTheme) GetUnitPath(unitId int32) string {
+	if mapping, ok := b.manifest.Units[unitId]; ok {
 		// Return relative path from theme base
 		return mapping.Image
 	}
 	return ""
 }
 
-func (b *BaseTheme) GetTilePath(terrainId int) string {
-	if mapping, ok := b.manifest.Terrains[int32(terrainId)]; ok {
+func (b *BaseTheme) GetTilePath(terrainId int32) string {
+	if mapping, ok := b.manifest.Terrains[terrainId]; ok {
 		// Return relative path from theme base
 		return mapping.Image
 	}
 	return ""
 }
 
-func (b *BaseTheme) IsCityTile(terrainId int) bool {
+func (b *BaseTheme) IsCityTile(terrainId int32) bool {
 	return IsCityTerrain(terrainId)
 }
 
-func (b *BaseTheme) IsNatureTile(terrainId int) bool {
+func (b *BaseTheme) IsNatureTile(terrainId int32) bool {
 	return IsNatureTerrain(terrainId)
 }
 
-func (b *BaseTheme) IsBridgeTile(terrainId int) bool {
+func (b *BaseTheme) IsBridgeTile(terrainId int32) bool {
 	return IsBridgeTerrain(terrainId)
 }
 
@@ -77,28 +77,28 @@ func (b *BaseTheme) GetThemeInfo() *v1.ThemeInfo {
 	return b.manifest.ThemeInfo
 }
 
-func (b *BaseTheme) GetAvailableUnits() []int {
-	units := make([]int, 0, len(b.manifest.Units))
+func (b *BaseTheme) GetAvailableUnits() []int32 {
+	units := make([]int32, 0, len(b.manifest.Units))
 	for id := range b.manifest.Units {
-		units = append(units, int(id))
+		units = append(units, id)
 	}
 	return units
 }
 
-func (b *BaseTheme) GetAvailableTerrains() []int {
-	terrains := make([]int, 0, len(b.manifest.Terrains))
+func (b *BaseTheme) GetAvailableTerrains() []int32 {
+	terrains := make([]int32, 0, len(b.manifest.Terrains))
 	for id := range b.manifest.Terrains {
-		terrains = append(terrains, int(id))
+		terrains = append(terrains, id)
 	}
 	return terrains
 }
 
-func (b *BaseTheme) HasUnit(unitId int) bool {
-	_, ok := b.manifest.Units[int32(unitId)]
+func (b *BaseTheme) HasUnit(unitId int32) bool {
+	_, ok := b.manifest.Units[unitId]
 	return ok
 }
 
-func (b *BaseTheme) HasTerrain(terrainId int) bool {
-	_, ok := b.manifest.Terrains[int32(terrainId)]
+func (b *BaseTheme) HasTerrain(terrainId int32) bool {
+	_, ok := b.manifest.Terrains[terrainId]
 	return ok
 }
